Feat/dyanmic channel game name var (#116)
* Fix: (`vertix-base`) - Skips checksum validation

Adds a command-line flag to skip checksum validation, allowing users to bypass the checksum check during configuration loading when needed.

* WIP on main

* Feat: (`vertix-bot`) - : add game var name to dynamic channel names

Adds the ability to include the user's current game name in dynamic channel names.

This allows for more informative and context-aware channel names, improving user experience.

- Introduces a `VAR_DYNAMIC_CHANNEL_GAME` constant for use in channel name templates.
- Fetches the user's current game from their presence.
- Updates the channel name assembly logic to include the game name if available.

* Refactor: (`vertix-bot`) - Extract getUserCurrentGame utility function

Moves the logic for fetching the user's current game from the DynamicChannelService to a new utility function in member.ts. This enhances code reusability and adheres to the Single Responsibility Principle. Additionally, updates the DynamicChannelService to utilize the new function for improved clarity and maintainability.

* Enhance: (`vertix-base`) - Introduce DEFAULT_DISCORD_MAX_CHANNEL_NAME_LENGTH constant

Adds a new constant for the maximum length of Discord channel names to improve code maintainability and adhere to the Single Responsibility Principle. Updates the DynamicChannelService to utilize this constant, replacing the hardcoded value for better reusability and clarity.